Extract rocket_deactivation() into inc/functions
Extraction of rocket_deactivation() from inc/main.php so we can test without loading main file and dealing with side-effects.
